Added PassengerSortingSystem

+++ Assets/ECS/Components/Car/CarComponent.cs
@@ -9,7 +9,7 @@ public struct CarComponent
 
     public Renderer renderer;
 
-    public float maxPassengersSlots;
+    public int maxPassengersSlots;
     public List<PassengerComponent> passengers;
     public List<PassengerComponent> reservedSeats;
